Added a static method ZipFile::openEmbeddedZipFile() that opens an embedded zipfile stored in the end of another file, only followed by a 4 byte integer (stored in the same byte order as the 4 byte unsigned integers in the zip header fields).
Added a program appendzip that appends a zipfile to another file and
writes the 4 byte integer indicating the start of the zipfile in the
end of the file.

Added a test_appzip test program.

Added an if construction in BackBuffer that tests if the VirtualSeeker
object it has in its possession is invalid. This test cannot be made
by the VirtualSeeker object, as it does not have access to the file it
is used for, to determine if it is invalid. Btw., it is invalid if
start_offset is a position after end_offset in the file, which happens
if start_offset+end_offset > true_file_length.