Added tests/Makefile.in, tests/Makefile.am and tests/.cvsignore.
Fixed a bug in tests/test_dircoll.cpp it was set to use a directory in
my home directory, which is unlikely to be present on other machines.

Finished reshuffling the source code. It is now organized like this:

zipios++/   contains the public headers.
src/        contains the implementation
tests/      contains the test and example programs

It compiles but there may be issues with stuff such as 'make
dist'. 'make doc' doesn't work right now.