Made sure all variable members get initialized.
Changed all tabs with 2 spaces and add "missing" { } to form blocks everywhere.
Removed m_stat_p because it is *dangerous* (since we have a chmod() it may become invalid).
Added a test in the release() function so we can catch any invalid reference counting.
Still need a solution for the stat(), chmod() and chown() which need to have their results checked.