Huge clean up so it compiles with most g++ warnings turned on and -std=g++11.
Moved doxyfile to the doc directory, made it a configurable file, upgraded to latest version of Doxygen (Ubuntu 14.04), and changed a few options.
Fixed some of the documentation, although at this point this was not my main focus.
Moved the appendzip tool to a new sub-directory called tools.
Added a .gitignore to hide a few files. You should not need that if you use cmake since all generated files end up in the destination directory.
Added necessary files to build with cmake. That should work for any Unices (including Linux and Mac OS/X) and MS-Windows.
Removed Makefile.vc files, .dsp/dsw files, win32 directory. The cmake output will work much better that those old things.
Removed zipios_defs.h (ended up empty!) and directly use [u]int32_t, [u]int16_t, and [u]int8_t as expected.
Removed the outputstringstream.h because with C++11 or better, stringstream cannot exist and thus we do not need to have a special class for that.
Removed the configure script usage from cmake so we are losing all the #ifdef, although the #ifdef are still in place (but need to be removed for most).
The dev/build script now understands a -m option to compile using cmake. It also runs the tests.
I added a dev/check to check recompile and run the test, but that was with the automake environment and I think I will only keep the build script.
Cleaned up the #include in most of the files to reduce them. Some may still be there commented out.
Made all enumerations use the class keyword and an explicit type.
Added dostime.c/h so we retrieve time as Unix time without having to know how to do the conversion everywhere it may be used.
Sorted the headers in the "right" order (if such there is) so a .cpp first includes its .h, then other library .h, then private .h, then system headers.
Started the redesign of exceptions which all derive from the std::runtime_error and are accessible by the entire library instead of just the collections.
As I saw them, made simple simplifications to the code (i.e. use 'auto' instead of a complicated iterator).
Noticed that the code was in need of some RAII and shared pointers. Will do that in a following wave of changes. Although I already have made fixes in a couple of files in that regard.
Fixed a bug in the SimpleSmartPointer, although we want to get rid of that class altogether and use shared pointers instead.
Added a few more initialization of variable members in the constructor list.
Removed all "using ...;" entries because that prevents (at least me) making sure that we know what the code really does.
Updated a couple of classes so they use m_... for their variable members (instead of nothing).
Updated the formatting (position of the {, position of the & for references, various alignments fixed because of tabs and such...)
Updated the copyright notices (1) I moved them at the top where I'm used to see them and (2) I updated the date to 2000-2015 and (3) I changed the "o/" with just "o" because some tools do not like it.
Updated the Makefile.am and other autogen files so it would take the changes in account and still compile the library, tool, tests. But these will be deleted very soon.
Moved the cppunit tests in the zipios_test namespace.
Added some "missing" copyright notices to various files.