More formatting changes and fixed a certain number of internal pointers using unique_ptr<>().
Added myself in the AUTHORS file. Clearly show where zipios++.spec now resides (although it was not updated to work with the cmake tools.)
Also properly configure zipios++.spec with a package name and version.
Added a couple of missing warnings (old style casts, no exception concerns.)
Added a dev/todo.sh script so we can see how many TODO, FIXME, etc. are defined in the code.
Moved some of the documentation to the .cpp files making the headers much less cluttered.
Improved readability of documentation that was worked on.
Removed the "go ahead delete that stream buffer" flags and the corresponding delete. We just cannot expect another class to handle our resources (too dangerous and it does not look like it was used anywhere).
Started adding m_ to all variable members, making it harder to mix members with locals or parameters.
Also marked global variables (static const, in general) with "g_" (i.e. global scope).
"Merged" the two "separator" declarations into one and named it "g_separator".
Replaced a certain number of zeroes (0) with nullptr when applicable.
Marked a certain number of internal points as std::unique_ptr<> since these are not shared.
Changed validity checks to use a virtual function so we do it in a single place instead of many many times.
Moved many declarations to the header which C++11 allows (i.e. "bool m_flag = false;") instead of the constructors.
Added some TODO and other notes about things that may need help (i.e. I fixed the VirtualSeeker 'int' type with 'off_t', but there are others that do not support 64 bit yet.)
Found out that the appendzip tool can be used along with a function available in the library so I make it public and installable.
Added exception named InvalidException for throwing when we detect invalid parameters calling a function.
Updated the README.
Removed the .cvsignore file.
Removed the autogen support in dev/build as we  are removing the corresponding scripts.
Changed the indentation to 4 spaces and various other style changes.
Removed many extraneous parenthesis making it harder to read expressions.
Moved many inline functions to their .cpp (created 2 new .cpp for that purpose). It generally is not useful and does not always make the code faster. Also it precludes updates to the code without also changing the interface.