Fixed up (see below) and wrote a full coverage test for FilePath using Catch.
Moved the documentation of the filepath.h to the filepath.cpp and fixed it and completed it.
Fixed some names so they make more sense.
Changed the prineTrailingSeparator() back to using an if() instead of a while().
Fixed the + operator which would not always append properly (i.e. it could transform a relative path to a full path which it should not in various cases.)
Added a cmake script to detect Catch, a C++ library to write tests.
Added a target in the makefile to run the Catch tests.
Converted the tests CMakeLists.txt to create a Catch test instead of the cppunit (or nothingness for other tests...)
Added a script to run the tests under coverage.
Added a script to fund all the TODO, XXX, FIXME and such comments.
Added a script to find unwanted spaces in the source files and fixed the last few as "required".