Changed the bare istream pointer returned by the getInputStream() functions to a shared pointer.
Moved the MatchName and MatchFileName classes to within the fcoll.cpp file. They are not needed in the header.
Moved the FileCollection documentation to the .cpp file.
Moved the isValid() inline function to the .cpp. In case we want to change its implementation it will make it easier.
Also moved the << operator of the FileCollection because it is actually quite big!