Started remodeling the FileEntry and FileCollection using shared pointers and moving things in base classes whenever possible and sensible.
Removed the CMAKE_MODULE_PATH from the dev/build script since it is now inside the CMakeLists.txt file.
Removed the CollectionCollection global instance because that would prevent the library from being reentrant (for threading support.)
Changed a set of size() function return type to size_t instead of int.
Moved a lot of the documentation to the FileEntry implementation file.
Transformed many definitions using a set of typedef's at the beginning of classes (i.e. shared pointers, buffers, vectors...)
Added a flag to know whether the CRC 32 was defined yet or not, also added an hasCrc() function.
Spent a bit more time on formatting as I find some more entries that were not updated yet.