Finished moving all the default functions from the derived classes to the FileEntry base class.
Also moved all the documentation from the various places were it was to the FileEntry implementation file (.cpp).
Renamed the uncompressed size field using the 'ed' in the name.
Added the mtime to a directory collection file using the mtime option.
Removed the std::ostream from a FileEntry pointer. It is not sensible to have it.
Made the time_t pointer in unix2dostime() a const.
Changed a few more uint32_t with size_t as it represents a size.
Made required updates to the various classes that derive from FileEntry.
Added a test so if someone tries to save a file that's too large in a zip file, an exception is raised.
Started adding override in a few places.