Updated the file names to full names.
Changed all the C++ headers to use .hpp instead of just .h.
Moved the zipheadio.hpp to the src/. directory since it should not be public. Also hid the dostime.h because it should not be necessary from the outside.
Updated the CMakeLists.txt and #include accordingly.
Removed the *.w32 and *.am exclusions from the zipios++ directory when installing since these do not exist anymore.
Added a TODO file and an mk script for completion.
Fixed up some formatting and documentation in the process.