Wrote test and fixed up the dostime.c/h files.
Started a test for the DirectoryEntry class.
Added support in the dostime.c file to avoid using invalid DOS dates. Instead we get 0 or -1 when a date cannot be represented.
Added a mindostime() and maxdostime() functions so we can initialize date/time values with such values.
Avoid an unnecessary call to localtime() just to get a struct tm pointer.
Also avoid a crash in case unix2dostime() is called with a null pointer.
Fixed the help screen of the dev/coverage script.
Renamed the BasicEntry as DirectoryEntry since that's the only way it gets used.
Changed the DirectoryEntry constructor to use a FilePath and comment only.
Removed the setName() function of the FileEntry class because it completely changes the validity of the object (and that was not taken in account).
Changed the FileEntry constructor to make use of FilePath instead of a direct string.
Actually wrote the hasCrc() function implementation.
Made the pruneTrailingSeparator() a global function only in the .cpp so it has nothing to do with the class declaration.
Removed the check_immediately parameter from the FilePath because it is not used and there is really no need to check anything immediately.
Changed the implementation to make use of an m_stat parameter instead of 6 predefined flags. That way we also have the size and mtime available.
Added the == operator to the FilePath.
Changed the FilePath::filename() function to directly return an std::string instead of another FilePath.
Removed all the friends from the various zip head classes. Make use of read() and write() functions instead of the << and >> operators.
Made the necessary modifications so it all compiles and works as expected.
Added test to cover the new additions to the FilePath.