Finished up the test that checks the DirectoryEntry. It is covered 100% from what I can tell.
Force the fileSize() to return 0 when a directory entry is created with a directory.
Documented the DirectoryEntry (most a copy from FileEntry at this point.)
Decided to change the unix2dostime() to use a value instead of a pointer.
Added a main() function so we can change the seed for rand() so each run is a bit different.