Went ahead and hid all the stream related headers with the source. ZipFile is what you are expected to use with the collections.
Fixed the doxy file to include the configured header and the .hpp[.in] files.
Fixed the Doxygen function reference, it uses \fn and not \func.
Found a copy of the Zip archive file format and added it to the project for quick reference.
Removed the getInputStream() with a FileEntry so we do not have to check whether the entry is indeed part of that collection.
Updated some documentation here and there (moved the mainpage to the zipfile.cpp file).
Broke out the EndOfCentralDirectory class to its own header file. Still need to break out the other two classes in two files with proper names.
Fixed the start/end offsets in the ZipFile constructor to use VirtualSeeker::offset_t instead of int.
Verified the DOS date and time, it can be read as one 32 bit value and then dealt with using dos2unixtime().
Made the writeCentralDirectory() function a static in the implemtation file so we do not have to have a declaration in the header.
Moved the example_zip.cpp from the tests to the tools to show it as a tool to list files in a zip archive. Also made it compile with the current version.
Added a new tool called dostime to convert numbers from Unix to DOS times and vice versa.
Added some more 'override' symbols.
Removed the need for the EndOfCentralDirectory class in the ZipFile.