Cleaned up the I/O from the ZipCDirEntry and ZipLocalEntry implementations.
Broke out the ziphead.hpp/cpp in various files (endofcentraldirectory/zipcdirentry/ziplocalentry).
Fixed the BackBuffer chunk_size parameter to use ssize_t instead of int.
Verify that streams are valid in many more places.
Replaced the various read and write functions in zipheadio.hpp with zipRead() and zipWrite() and removed the endian low level swap functions (they are not necessary if you read from a buffer you can also make sure you read the value as little or big endian as you desire).
Replaced the operator == of the ZipLocalEntry with an isEqual() function, which seems cleaner at this point.
Removed one of the getInputStream() because it called the other with entry->getName() and it can be confusing if you have to collections and use an entry from the other and it works...
Started to work on a fix to get the file offset in the FileEntry instead of the ZipCDirEntry because that way we do not have to overload the getInputStream() function.
Made the offset_t part of the configuration file instead of the VirtualSeeker so that way we can reuse it in various places.
Changed the putNextEntry() function signature to make use of a shared pointer as expected by the m_entries vector.
Fixed a couple of << and >> that were still used instead of .read() and .write().
Added a test to verify that the += operator works as expected on vectors.
Changed the directory entry test so it shows 100% coverage.
With all those changes a couple of unit tests do not cover 100% of the code anymore...
Removed some more explicit on constructors. I do not think it is as necessary with C++11.
Dropped the meta-iostreams.hpp because really it served no real purpose