Added --durations to coverage tests when generating output for website.
Fixed tests so the --full command line option works as expected and publish on a local site.
Added tests to verify the zipRead() and zipWrite() functions.
Fixed the directory entry tests so they don't take forever to run.
Show an extra help for the --seed command line option of the tests.