Cleaned up the #include's and fixed the \file documentation in entire project.
Removed all the \anchor and \ref because there is no way I would maintain those. Just write the correct class name and you'll get the correct link.
Removed the const in the return type of the operator = wherever it was used.
Added script to recompile the documentation without having the re-run dev/build (see mk in root folder).
Added script to publicly publish documentation if dev/coverage works.
Added a link in the index.php to the documentation.
Applied a better fix for the BINARY_DIR in zipios.doxy.in (this time it works).
Added support for \TODO and \FIXME in source to work as \todo.
Added path and pattern to get the example to show up in the documentation.
Fixed various things in the documentation, added some of the missing documentation.
Turned on MATHJAX although we don't currently use it.
Delete the copy contructor and operator from the FilterOutputStreambuf (like in the FilterInputStreambuf).
Fixed many more 'int' to use the proper value so it will work in 64 bit mode.
Removed the copy operators of the various FileEntry because the default will work as expected.
Properly check the size of the entry offset (now that it can be more than 32 bits).
Verify that the filename, extra buffer, and comment are not too large before doing a write().
Removed the dynamic_cast<>() from the zipfile and various buffer implementation as I extended the FileEntry class to work as expected there too.
Finally could remove the ZipCDirEntry predefinition from the ZipLocalEntry.hpp header.