Added the DirectoryCollection unit test and that is covered 100%.
Fixed the getFileName() so it returns the last segment whether the FileEntry is a directory or not.
Added a comment line option to name the test to run with dev/check.
Fixed a few '++' that were after instead of before the variable.
Removed the load_now variable from the DirectoryCollection, it was not used anywhere and there is no reason for having it.
Better clean up a DirectoryCollection on a close.
Removed the various optimizations that would try to avoid loading the DirectoryCollection once needed.
Fully documented the DirectoryCollection and FileCollection classes.
Fixed the DirectoryCollection so it includes the directories that it finds (including the root directory.) This fixes the count and that way we can be sure to save those directories in Zip archives.
Applied various optimizations to the code.