Applied a fix to the matchEntry() in the CollectionCollection implementation as using the iterator somehow would crash. I use the smart pointer instead now.
Added a specialized mustBeValid() function to the CollectionCollection since any collection may become invalid.
Added a clear() of the collections in the close() of the CollectionCollectin to save memory.
Verify that cloning a collection returns the correct type of pointer.
Completed the CollectionCollection test by adding a test of all the files in a "complex" tree of collections.
Verify that a collection cannot be added to itself.