Many changes to the EndOfCentralDirectory class to use proper sizes and offsets.
Also removed the disk number from being a member since we never use those numbers.
Removed the "total entries" since we expect that number to be the same as "entries".
Renamed the "cdir" part of a variable member "central_directory" to be clear.
Fixed the function names to match the natural convention of the project.
Fixed the read and write accordingly.
Added checks in the write function in case of overflows.
Added documentation to the EndOfCentralDirectory class.
Applied an optimization in a loop.
Continued work on testing the EndOfCentralDirectory and such.