Actually added all the documentation in the EndOfCentralDirectory class.
Noticed that the test in ZipFile is incorrect. The old test expected the end of the Central Directory to match the beginning of the EndOfCentralDirectory, which is wrong. It does not need to be contiguous. The new test makes use of the Central Directory size to verify we are at the end once we read it.
Removed the test on the output stream in the write() function. It is not sensible. Instead we will throw if the output stream is not valid.
Did some clean up of the #include's as I made more changes with SunOS there were a few duplicates.
Fixed the getOffset() return value, was using size_t instead of offset_t.
Compeletely removed the m_eocd_offset_from_end since we are not supposed to test that value, ever.
Finally decided to move the m_filename = ... to the FileCollection instead of the ZipFile.
Defined the default filename for a collection as a global variable so we can reuse the exact same string multiple times.