Finished up the reading a zipfile test with about 100% coverage (hard to tell for sure since we still have to test the write feature and we may notice a few other things...)
Removed a few more tests on the input stream, if invalid, it will throw anyway.
Removed all those functions in link with the getNextEntry() which was a la Java but never used properly by the rest of the API.
Moved the getNextEntry() code to the constructor since there is no need to have another function for it. Plus, once we remove all the extras it really isn't so bad.
Changed the ZipInputStreambuf::underflow() implementation to better show that we have one section per entry compression method.
Added a test to avoid accessing the input stream in case a buffer size is zero (in zipRead() of buffers and strings.)
Added many tests to make sure we check every single line of code. We still need more tests to ensure all cases are considered, though.