Finishing up with the reading of a zip file.
Fixed various little things (i.e. Z_OK instead of 0, Bytef instead of unsigned char.)
Changed the initialization of the ZipCentralDirectoryEntry to better match what's used (started on the output feature.)
Changed the handling of the isDirectory() so it works properly for Zip file entries (i.e. we cannot check the disk for those!)
Fixed the output of directories which need to include a "/" at the end of the filename when saving to a header.
Changed the toString() of a couple of entries to output (directory) for directory entries.
Fixed the trailing data descriptor test, it is bit 3 and thus (1 << 3). The old code used 4 instead.
Changed the size test in setExtra() to allow any 16 bit size.
Added more documentation of private classes.
Removed the putNextEntry() using a filename because that won't work to add meta data such as comments and extra buffers.
Test the << operator with DirectoryCollection class.
Added a test with a zip buffer that has invalid data to test that reading that data fails as expected.