Implemened the necessary code to make the output work.
Removed yet another "let's init now" flag.
Moved the compression method/level setup to the base FileEntry class so that way it is available for the zip output stream.
Moved the Comment to the FileEntry so that way it can be easily shared between directory and zip local entry (the copy works a lot better).
Strengthen the DeflateOutputStreambuf class so we can actually save STORED files and directories (we could not before, all were DEFLATED and had to be regular files.)
Defined a new compression level which goes from 0 to 100 so it can be applied to any kind of compression library.
Renamed the function getCount() into getSize() since it is the size of the uncompressed file.
Updated the GZIP files accordingly so it still compiles.
Documented the DeflateOutputStreambuf class.