Moved the extra field buffer from the ZipLocalEntry to the FileEntry so that way it is available in all circumstances.
Added a TODO entry.
Fixed the DeflateOutputStreambuf::sync() function, it is supposed to return -1 as per spec.
Transformed an error being printed in std::cerr into a throw. If the zlib library fails, the output may not be consistent.
Deleted some copy constructor and assignment operators.
Made a DirectoryCollection valid if it only includes a regular file.
Added support to create a Zip archive with just a regular file in the DirectoryCollection.
Changed the validity of a FileCollection from invalid when empty (on creation) to valid.
Fixed the saveCollectionToArchive() to properly handle exceptions by marking the output stream as failed if one happens.
Fixed a terminate() call by calling the finish() function on the output stream defined in saveCollectionToArchive() before the destructor gets called.
Cleaned up the ZipLocalEntry constructor as it was never called with a filename and extra field buffer.
Added tests with one regular file in a Zip archive. And one with a comment that's too long.
Updated the tests accordingly.
Enhanced, added more documentation.