Started adding tests for compression levels and a few other things.
Added --seed to the dev/check script so I can check the same case over and over again.
Applied a simplification to avoid many calls in case the overflow() is called with an empty input buffer.
Changed the toString() design so only the FileEntry has a definition. After all, the data used in the string is available at that level.
Made the FileCollection writable to where one can add a FileEntry to it.
Applied a fix to the ZipFile headers in case the compression level is COMPRESSION_LEVEL_NONE.
Started adding tests changing the level of compression. This version fails with files over 32Kb.
Added tests of an invalid (too long) comments, extra buffers, and filenames.