Fixed the `dev/upload.website` script, it copies the images as expected and uses the right path for the documentation.
Updated the `index.html`, especially the logo URL, the versions, links to github for support & source code.
Added links to zipios v2.2 and v2.1 documentations.
Added a link to launchpad where a debian (Ubuntu 16.04 at the moment) can be downloaded.
Fixed the footer so the copyright notice is clearly visible even with the gradient.
Updated the `dev/check` script with the new DOS Date & Time test names.
Updated the DOS Date & Time test to have names that match the other tests.