Verify the version in the tests.
Updated the coverage so it will properly test the version.
This is to make a little more sure that whenever I update the changelog, the
tests tell me to also fix the CMakeLists.txt version. They need to be in sync.
Added a --source-path command line option to the tests so they can run the
dev/version script as expected.