Made the modern coverage test pass.
The coverage detects when a test does not ever use a CATCH_REQUIRE() call.
We had several such tests. These are fixed.

Fixed the DOS time minimum test so it actually does what it is supposed to
do (i.e. it assumed that the minimum date was 0, which as a Unix timestamp
is not the case).

One test would do a chdir() to a folder we then deleted, creating all sorts
of problems and failing the test.

Cleaned up tests further so we have more sections with the proper naming
conventions.

Updated docs about the coverage and the variable name used to turn on the
coverage (removed the name of the project, not necessary and not compatible
with our main coverage script).